Improvements to ViewJson and ViewMarkup feature
